Test Drive Fixes

+++ src/main/java/org/moonlightcontroller/managers/models/messages/Alert.java
@@ -4,19 +4,19 @@ import java.util.List;
 
 public class Alert extends Message {
 
-	private int origin_dpid;
+	private long origin_dpid;
 	private List<AlertMessage> messages;
 	
 	// Default constructor to support Jersy
 	public Alert() {}
 	
-	public Alert(int xid, int origin_dpid, List<AlertMessage> messages) {
+	public Alert(int xid, long origin_dpid, List<AlertMessage> messages) {
 		super(xid);
 		this.origin_dpid = origin_dpid;
 		this.messages = messages;
 	}
  	
-	public int getOrigin_dpid() {
+	public long getOrigin_dpid() {
 		return origin_dpid;
 	}
